requirements.txt
colorama==0.3.9
Django==2.0.4
django-pandas==0.5.1
ipdb==0.11
ipython==6.3.1
mysqlclient==1.3.12
python-dateutil==2.7.2
tabulate==0.8.2
